Make the block size floating-point.
This allows more precisely controlling how the mosaic aligns, like the Unity
version does.  The edges of the mosaic won't antialias for fractional values.

+++ mosaix-commandline/CommandLine.cpp
@@ -51,7 +51,7 @@ int main(int argc, char *argv[])
             break;
 
         case 'b':
-            options.block_size = atoi(optarg);
+            options.block_size = (float) atof(optarg);
 
             if(options.block_size == 0)
             {
